backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
mysql-connector-python==8.2.0
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Werkzeug==2.3.7
eventlet==0.33.3
redis==5.0.1